pyproject.toml
[project]
name = "ai-webpage-summarizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "ollama>=0.6.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
